removed ponintless ENABLE_PORT_SECURITY define & reorganized config loading

+++ src/common/singleton.h
@@ -11,8 +11,8 @@ template <typename T> class singleton
 		singleton()
 		{ 
 			assert(!ms_singleton);
-			long offset = (long) (T*) 1 - (long) (singleton <T>*) (T*) 1; 
-			ms_singleton = (T*) ((long) this + offset);
+			intptr_t offset = (intptr_t) (T*) 1 - (intptr_t) (singleton <T>*) (T*) 1;
+			ms_singleton = (T*) ((intptr_t) this + offset);
 		} 
 
 		virtual ~singleton()
